Ignore systemd errors when force overwriting existing deployments
It is possible to have a partially installed previous deployment
where keycloak_jboss_home exists, but the keycloak service is not
set up in systemd.  In a forced overwrite installation, this causes
a systemd failure when we try to stop the old running deployment.
We can safely ignore this and continue with the cleanup.