Fix lint issues
This fixes lint issues detected by yamllint and ansible-lint.  The
type of issues fixed are tasks with missing names, improper indentation,
line-length, and skipping tasks that use the command module with the
newer argv format (which ansible-lint doesn't handle).

Note that there are still some issues that yamllint will report about
truthy values.  These failures will be ignored in yamllint configuration
that will be added in a later commit that adds the CI tests.