Update java dependency to openjdk 1.8.0 headless
There are some challenges running keycloak on newer versions of
Java such as Java 11.  It is possible to get it to work with some
extra options that need to be passed to java.  We will need to
ultimately use some fact-based logic to detect what version of
java is installed for a particular target system, then adjust the
options we pass to java accordingly.  For now, this just pins us
to using OpenJDK 1.8.0, which will still be supported for some
time.  This also uses the headless version of java to cut down on
the number of packages dependencies we need to install on the
target system.