Detect and use python3 if available
On target systems where python3 is available, we want to  prefer
its use over python2.  In the case of modern Fedora versions, some
of the required packages needed for this role are no longer provided
for python2 (such as python2-firewall).  This forces the use of python3
for configuring the firewall, so we should just set the interpreter
for the user instead of requiring them to figure out why it fails and
then correctig it themselves.

This was tested agaist CentOS7, Fedora28, and Fedora29 using the
soon to be commited tests.