Add molecule tests
This adds linting and functional tests for the keycloak role.  We
use Molecule for the tests, which provides provisioned containers
that can be used to test the role against different target plarforms.
Detaild on running the tests are in the README.  The tests are
currently passing on the latest CentOS 7, Fedora 28, and Fedora 29
images.

It is planned to use these tests for CI in the near future.