Restrict CI for branch pushes
This restricts pushes to branches to only trigger CI on master.  We
will run CI on all PRs still.  This prevents running the CI twice
when I push a topic branch that will be used for a PR.  If we start
using branches for other purposes, we can whitelist CI for those
branches as needed.