Update testing doc, specifically address role name issue
I had a lot of trouble trying to run the molecule tests in my local
repo. This commit adds documentation to the README.md to cover the
various gotchas I encountered with the hope others will benefit.

Some fixes were made the markdown formatting as well, mostly with the
use of section titles and block quotes.

Signed-off-by: John Dennis <jdennis@redhat.com>