Allow firewall configuration to be optionally skipped
In some cases, one might want to skip having the role configure
the firewall.  For example, if the firewall is already disabled
and other network services are running on the system, having this
role enable and configure the firewall will result in non-keycloak
ports being blocked.

This adds a new optional parameter to skip firewall configuration.
The default behavior is to configure the firewall, which retains
the previous behavior for backwards compatibility.