install pyOpenSSL package more dynamically
RHEL8 cannot install the python3 version of pyOpenSSL with that name.
You must use "python3-pyOpenSSL" to install for RHEL8.

Adding a fact to python_2_3_test.yml to define which package name to
use for pyOpenSSL:  keycloak_dep_pyOpenSSL

Changing package install for dependencies to use the fact that we set
in the python_2_3_test.yml task.

Resolves: issue #24

https://github.com/nkinder/ansible-keycloak/issues/24