fix: autodetection of python major version
This patch adds provisions so the python major version is handled
uniformly, autodetecting the version only if ansible_python_interpreter
is not defined at the time the role is applied.

Whether the interpreter is discovered or set previously, an extra
variable with the major Python version in use is defined, so further
conditional blocks can take actions, like accounting for different Python
library names for example.